requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
distlib==0.3.7
exceptiongroup==1.1.2
filelock==3.12.2
frozenlist==1.4.0
identify==2.5.26
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
nodeenv==1.8.0
packaging==23.1
platformdirs==3.9.1
pluggy==1.2.0
pre-commit==3.3.3
PyJWT==2.8.0
pytest==7.4.0
pytest-asyncio==0.21.1
PyYAML==6.0.1
requests==2.31.0
tomli==2.0.1
urllib3==2.0.4
virtualenv==20.24.2
yarl==1.9.2
